requirements.txt
Django==4.2.22
dj-database-url==2.3.0
python-decouple==3.8
Pillow==11.1.0
whitenoise==6.9.0
